made the interface for raytraced procedural geometry a bit more flexible
